MHDapp/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==2.3.0
numpy==2.2.4
packaging==24.1
pandas==2.2.3
pillow==11.1.0
plotly==6.3.1
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
win32_setctime==1.2.0
xarray==2025.10.1
